pyproject.toml
[build-system]
requires = ["setuptools >= 64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "countryguessr"
description = "Code base necessary to reproduce the results from the paper 'Exercising Good Practices of Machine Learning Research: A Case Study of Environment Image Classification'"
readme = "README.md"
authors = [
  {name = "Lourenço A. Rodrigues", email = "lourenco@uni-bremen.de"},
  {name = "Kieran Taylor", email = "taylor@uni-bremen.de"},
  {name = "Jordan Behrendt", email = "jobe@uni-bremen.de"},
  {name = "Leon Brenig", email = "brenig@uni-bremen.de"},
]
license = {text = "MIT License"}
dependencies = [
    "geopandas==0.13.2",
    "matplotlib==3.7.3",
    "numpy==1.24.3",
    "openai-clip==1.0.1",
    "pandas==2.0.3",
    "pypdf2==3.0.1",
    "pyyaml==6.0.1",
    "reportlab==4.0.7",
    "scikit-learn==1.3.2",
    "scipy==1.10.1",
    "seaborn==0.12.2",
    "shapely==2.0.2",
    "tensorflow==2.13.1",
    "torch==2.1.0",
    "tqdm==4.66.1",
    "requests==2.31.0",
    "selenium==4.16.0",
    "leafmap==0.35.11",
    "fiona<1.10.0",
    "webdriver-manager==4.0.1"
]
classifiers = [
    "Intended Audience :: Researchers",
    "Topic :: Machine Learning :: Research",
    "Topic :: Machine Learning :: Geolocation",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
]
dynamic = ["version"]


[tool.setuptools.dynamic]
version = {attr = "countryguessr.__version__"}

[project.urls]
Repository = "https://github.com/jordanbehrendt/good_practices_ml"
